🗺️ Atlas: Splitting Query Planner and Executor Blobs
Refactored `src/query/planner/mod.rs` and `src/query/executor/mod.rs` by extracting their test modules into separate `tests.rs` files to resolve "Blob" anti-pattern and improve navigability.
